Code clean-up, linting issue fix
